+++ CHANGELOG.md
@@ -44,7 +44,7 @@ interfaces.
 * Removed "functions.php" so that /* Replaced /* Replaced /* Replaced Guzzle */ */ */ is truly PSR-4 compliant. These
   functions are now implemented in `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\Utils` using camelCase.
   `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\json_decode` moved to `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\Utils::jsonDecode`.
-  `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\batch` moved to `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\Utils::batch`.
+  `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\batch` moved to `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\Pool::batch`.
   `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\get_path` moved to `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\Utils::getPath`.
   `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\set_path` moved to `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\Utils::setPath`.
 * Removed the "headers" event. This event was only useful for changing the
@@ -66,6 +66,10 @@ interfaces.
   option name mapping to a function that applies the option value to a request.
 * Removed `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\ClientInterface::sendAll` and marked
   `/* Replaced /* Replaced /* Replaced Guzzle */ */ */Http\Client::sendAll` as deprecated.
+* Removed the concept of "throwImmediately" from exceptions and error events.
+  This control mechanism was used to stop a transfer of parallel requests from
+  completing. This can now be handled by throwing the exception or by
+  cancelling a pool of requests or each outstanding future request individually.
 
 4.2.2 (2014-09-08)
 ------------------